remove some unnessasary files
